requirements.txt
Appium-Python-Client==3.1.0
selenium==4.15.2
pandas==2.1.4
openpyxl==3.1.2
